Updated documentation and added some debugging code for building on win
