+++ docs/source/api/algorithms/builders.rst
@@ -2,7 +2,13 @@
 .. Distributed under the terms of the Apache License 2.0
 
 #####################################
-Angles
+Molecular and System Builders
 #####################################
-.. automodule:: exatomic.algorithms.angles
+.. automodule:: exatomic.util.builder
+    :members:
+
+.. automodule:: exatomic.algorithms.geometry
+    :members:
+
+.. automodule:: exatomic.algorithms.packing
     :members:
