+++ docs/source/api/algorithms/orbital.rst
@@ -1,0 +1,11 @@
+.. Copyright (c) 2015-2018, Exa Analytics Development Team
+.. Distributed under the terms of the Apache License 2.0
+
+#####################################
+Orbital Algorithms
+#####################################
+.. automodule:: exatomic.algorithms.orbital
+    :members:
+
+.. automodule:: exatomic.algorithms.orbital_util
+    :members:
