+++ docs/source/api/algorithms/structural.rst
@@ -1,0 +1,20 @@
+.. Copyright (c) 2015-2018, Exa Analytics Development Team
+.. Distributed under the terms of the Apache License 2.0
+
+#####################################
+Structural Analysis
+#####################################
+.. automodule:: exatomic.algorithms.angles
+    :members:
+
+.. automodule:: exatomic.algorithms.distance
+    :members:
+
+.. automodule:: exatomic.algorithms.pcf
+    :members:
+
+.. automodule:: exatomic.algorithms.displacement
+    :members:
+
+.. automodule:: exatomic.algorithms.diffusion
+    :members:
