+++ docs/source/api/error.rst
@@ -1,0 +1,8 @@
+.. Copyright (c) 2015-2018, Exa Analytics Development Team
+.. Distributed under the terms of the Apache License 2.0
+
+#####################################
+Core Exceptions
+#####################################
+.. automodule:: exatomic.core.error
+    :members:
