+++ docs/source/api/tests/algorithms_tests.rst
@@ -1,0 +1,20 @@
+.. Copyright (c) 2015-2018, Exa Analytics Development Team
+.. Distributed under the terms of the Apache License 2.0
+
+#####################################
+Algorithms Tests
+#####################################
+.. automodule:: exatomic.algorithms.tests.test_basis
+    :members:
+
+.. automodule:: exatomic.algorithms.tests.test_distance
+    :members:
+
+.. automodule:: exatomic.algorithms.tests.test_numerical
+    :members:
+
+.. automodule:: exatomic.algorithms.tests.test_orbital
+    :members:
+
+.. automodule:: exatomic.algorithms.tests.test_overlap
+    :members:
