+++ docs/source/api/tests/base_tests.rst
@@ -1,0 +1,11 @@
+.. Copyright (c) 2015-2018, Exa Analytics Development Team
+.. Distributed under the terms of the Apache License 2.0
+
+#####################################
+Base Tests
+#####################################
+.. automodule:: exatomic.interfaces.tests.test_cube
+    :members:
+
+.. automodule:: exatomic.interfaces.tests.test_xyz
+    :members:
