+++ docs/source/api/widgets.rst
@@ -1,0 +1,17 @@
+.. Copyright (c) 2015-2018, Exa Analytics Development Team
+.. Distributed under the terms of the Apache License 2.0
+
+#############################
+Visualization
+#############################
+.. automodule:: exatomic.widgets.widget
+    :members:
+
+.. automodule:: exatomic.widgets.traits
+    :members:
+
+.. automodule:: exatomic.widgets.widget_base
+    :members:
+
+.. automodule:: exatomic.widgets.widget_utils
+    :members:
