+++ docs/source/index.rst
@@ -27,6 +27,11 @@ method to get started with visualization is through the `Jupyter notebook`_.
     :caption: General
 
     api/index.rst
+
+.. toctree::
+    :maxdepth: 2
+    :caption: Examples
+
     notebooks/nwchem.ipynb
     notebooks/qe.ipynb
 
@@ -34,8 +39,12 @@ method to get started with visualization is through the `Jupyter notebook`_.
     :maxdepth: 3
     :caption: Software
 
-    software/adf/index.rst
-    software/nwchem/index.rst
+    software/adf.rst
+    software/gaussian.rst
+    software/molcas.rst
+    software/nbo.rst
+    software/nwchem.rst
+    software/qe.rst
 
 .. toctree::
     :maxdepth: 1
