+++ docs/source/software/adf.rst
@@ -1,0 +1,24 @@
+.. Copyright (c) 2015-2018, Exa Analytics Development Team
+.. Distributed under the terms of the Apache License 2.0
+
+.. _adf-label:
+
+#############################
+`ADF`_
+#############################
+.. automodule:: exatomic.adf.editor
+    :members:
+
+.. automodule:: exatomic.adf.inputs
+    :members:
+
+.. automodule:: exatomic.adf.output
+    :members:
+
+.. automodule:: exatomic.adf.tests.test_editor
+    :members:
+
+.. automodule:: exatomic.adf.tests.test_orbital
+    :members:
+
+.. _ADF: https://www.scm.com/
