--- docs/source/software/adf/index.rst
@@ -1,11 +1,0 @@
-.. Copyright (c) 2015-2018, Exa Analytics Development Team
-.. Distributed under the terms of the Apache License 2.0
-
-.. _adf-label:
-
-#############################
-`ADF`_
-#############################
-
-
-.. _ADF: https://www.scm.com/
