+++ docs/source/software/gaussian.rst
@@ -1,0 +1,27 @@
+.. Copyright (c) 2015-2018, Exa Analytics Development Team
+.. Distributed under the terms of the Apache License 2.0
+
+.. _gaussian-label:
+
+#############################
+`Gaussian`_
+#############################
+.. automodule:: exatomic.gaussian.editor
+    :members:
+
+.. automodule:: exatomic.gaussian.inputs
+    :members:
+
+.. automodule:: exatomic.gaussian.output
+    :members:
+
+.. automodule:: exatomic.gaussian.tests.test_editor
+    :members:
+
+.. automodule:: exatomic.gaussian.tests.test_inputs
+    :members:
+
+.. automodule:: exatomic.gaussian.tests.test_output
+    :members:
+
+.. _Gaussian: http://gaussian.com/
