+++ docs/source/software/molcas.rst
@@ -1,0 +1,27 @@
+.. Copyright (c) 2015-2018, Exa Analytics Development Team
+.. Distributed under the terms of the Apache License 2.0
+
+.. _molcas-label:
+
+#############################
+`Molcas`_
+#############################
+.. automodule:: exatomic.molcas.editor
+    :members:
+
+.. automodule:: exatomic.molcas.inputs
+    :members:
+
+.. automodule:: exatomic.molcas.output
+    :members:
+
+.. automodule:: exatomic.molcas.tests.test_editor
+    :members:
+
+.. automodule:: exatomic.molcas.tests.test_inputs
+    :members:
+
+.. automodule:: exatomic.molcas.tests.test_output
+    :members:
+
+.. _Molcas: https://gitlab.com/Molcas
