+++ docs/source/software/nbo.rst
@@ -1,0 +1,18 @@
+.. Copyright (c) 2015-2018, Exa Analytics Development Team
+.. Distributed under the terms of the Apache License 2.0
+
+.. _nbo-label:
+
+#############################
+`NBO`_
+#############################
+.. automodule:: exatomic.nbo.editor
+    :members:
+
+.. automodule:: exatomic.nbo.inputs
+    :members:
+
+.. automodule:: exatomic.nbo.output
+    :members:
+
+.. _NBO: http://nbo6.chem.wisc.edu/webnbo_css.htm
