+++ docs/source/software/nwchem.rst
@@ -1,0 +1,22 @@
+.. Copyright (c) 2015-2018, Exa Analytics Development Team
+.. Distributed under the terms of the Apache License 2.0
+
+.. _nwchem-label:
+
+#############################
+`NWChem`_
+#############################
+.. automodule:: exatomic.nwchem.basis
+    :members:
+
+.. automodule:: exatomic.nwchem.editor
+    :members:
+
+.. automodule:: exatomic.nwchem.inputs
+    :members:
+
+.. automodule:: exatomic.nwchem.output
+    :members:
+
+
+.. _NWChem: https://github.com/nwchemgit/nwchem
