--- docs/source/software/nwchem/index.rst
@@ -1,11 +1,0 @@
-.. Copyright (c) 2015-2018, Exa Analytics Development Team
-.. Distributed under the terms of the Apache License 2.0
-
-.. _nwchem-label:
-
-#############################
-`NWChem`_
-#############################
-
-
-.. _NWChem: http://www.nwchem-sw.org/index.php/Main_Page
