+++ docs/source/software/qe.rst
@@ -1,0 +1,51 @@
+.. Copyright (c) 2015-2018, Exa Analytics Development Team
+.. Distributed under the terms of the Apache License 2.0
+
+.. _qe-label:
+
+#############################
+`Quantum ESPRESSO`_
+#############################
+.. automodule:: exatomic.qe.pw.input
+    :members:
+
+.. automodule:: exatomic.qe.pw.output
+    :members:
+
+.. automodule:: exatomic.qe.cp.dynamics
+    :members:
+
+.. automodule:: exatomic.qe.psp.jobfile
+    :members:
+
+.. automodule:: exatomic.qe.psp.upf
+    :members:
+
+.. automodule:: exatomic.qe.pw.parsers
+    :members:
+
+.. automodule:: exatomic.qe.pw.composer
+    :members:
+
+.. automodule:: exatomic.qe.pw.header
+    :members:
+
+.. automodule:: exatomic.qe.pw.footer
+    :members:
+
+.. automodule:: exatomic.qe.pw.scf
+    :members:
+
+.. automodule:: exatomic.qe.pw.classes
+    :members:
+
+.. automodule:: exatomic.qe.core.base
+    :members:
+
+.. automodule:: exatomic.qe.core.atomic_positions
+    :members:
+
+.. automodule:: exatomic.qe.types
+    :members:
+
+.. _Quantum ESPRESSO: https://www.quantum-espresso.org/
