+++ exatomic/algorithms/delocalization.py
@@ -14,10 +14,10 @@ import os
 import numpy as np
 import pandas as pd
 import seaborn as sns
-from exa.mpl import _gen_figure
+from exa.util.mpl import _gen_figure
 from exa.util.units import Energy
 from exatomic import gaussian, nwchem
-from exatomic.mpl import plot_j2_surface, plot_j2_contour
+#from exatomic.mpl import plot_j2_surface, plot_j2_contour
 
 sns.mpl.pyplot.rcParams.update({'text.usetex': True,
                                 'font.family': 'serif',
