+++ exatomic/algorithms/numerical.py
@@ -293,7 +293,5 @@ class Shell(object):
         self.rs = rs
         self.ns = ns
 
-try:
-    shell_type.define(Shell.class_type.instance_type)
-except:
-    pass
+print(dir(Shell))
+shell_type.define(Shell.class_type.instance_type)
