+++ exatomic/qe/core/base.py
@@ -6,7 +6,8 @@ Base Output Parser
 ###############################
 """
 import pandas as pd
-from exa import Parser, Typed
+from exa import Parser
+from exa.typed import Typed
 from exatomic.core.atom import Atom
 from .atomic_positions import AtomicPositions
