--- exatomic/tests/__init__.py
@@ -1,3 +1,0 @@
-## -*- coding: utf-8 -*-
-## Copyright (c) 2015-2018, Exa Analytics Development Team
-## Distributed under the terms of the Apache License 2.0
