--- exatomic/tests/test_atom.py
@@ -1,21 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (c) 2015-2018, Exa Analytics Development Team
-## Distributed under the terms of the Apache License 2.0
-#"""
-#Tests for the Atom DataFrame
-##############################
-#The tests here use some contrived examples.
-#"""
-#import numpy as np
-#from unittest import TestCase
-#from exa import units
-#from exa.core.dataframe import ColumnError
-#from exatomic.atom import Atom
-#
-#
-#class TestAtom(TestCase):
-#    """Tests for :class:`~exatomic.atom.Atom`."""
-#    def test_init(self):
-#        """Test that the atom dataframe raises errors correctly."""
-#        with self.assertRaises(ColumnError):
-#            Atom()
